feat: Implement Solution.canChoose function
This commit adds the implementation of the `canChoose` function in the `Solution` class. This function checks if it is possible to choose n disjoint subarrays from the array nums such that the i^th subarray is equal to groups[i] and if i > 0, the (i-1)^th subarray appears before the i^th subarray in nums. It iterates through the groups and checks if each group can be found in nums in the correct order. If any group cannot be found, it returns False. Otherwise, it returns True.